requirements.txt
streamlit
python-dotenv
sentence-transformers
faiss-cpu
openai
google-generativeai
requests
numpy